requirements.txt
kafka-python==2.0.2
requests==2.31.0
pyspark==3.4.0
pymongo==4.6.1
numpy>=1.21.0
pandas>=1.3.0
apache-airflow==2.8.1
pytest==7.4.0
